pyproject.toml
[project]
name = "ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.21.0",
    "openai-agents>=0.9.0",
    "python-dotenv>=1.2.1",
]
